Implement dynamic allocation for PStateStack Реализовал динамическое расширение стека якорей

+++ src/internals/parser/helpers/tbuilder.h
@@ -6,8 +6,8 @@
 
 typedef struct
 {
-    size_t cp;
-    size_t cap;
+    int64_t cp;
+    int64_t cap;
     TNode** anchors;
 } PStateStack;
